Ordenar archivos y rutas
Se orden las archivos y se agregan las rutas para login, signup, y diary
