+++ src/components/Header/NavigationHeader.js
@@ -1,17 +1,14 @@
 import { useSelector } from 'react-redux';
 import { useLocation } from 'react-router-dom';
 import { selectUser } from 'redux/auth/selectors';
-import { getSelectedDate } from 'redux/date/selectors';
 import { selectCalculateValue } from 'redux/calculate/selectors';
-import { initialDate } from 'helpers/constants';
-
 import { Link, HeaderNavigationWrapper } from './NavigationStyledHeader';
 
 export const NavigationHeader = ({ isLoggedIn }) => {
   const location = useLocation();
   const { data } = useSelector(selectUser);
   const { formData } = useSelector(selectCalculateValue);
-  const selectedDate = useSelector(getSelectedDate);
+
 
   return (
     <HeaderNavigationWrapper
@@ -21,13 +18,7 @@ export const NavigationHeader = ({ isLoggedIn }) => {
       {isLoggedIn ? (
         (data?.height || formData?.height) && (
           <>
-            <Link
-              to={`/diary/${
-                selectedDate ? selectedDate.split('.').join('-') : initialDate
-              }`}
-            >
-              DIARIO
-            </Link>
+            <Link to="/diary">DIARIO</Link>
             <Link to="/calculate">CALCULADORA</Link>
           </>
         )
