--- src/components/Layout.jsx
@@ -1,18 +1,0 @@
-import { Suspense } from 'react';
-import NavBar from './NavBar';
-import { Outlet } from 'react-router-dom';
-
-const Layout = () => {
-  return (
-    <>
-      <NavBar />
-      <div>
-        <Suspense>
-          <Outlet />
-        </Suspense>
-      </div>
-    </>
-  );
-};
-
-export default Layout;
