+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
-import { store } from './app/store';
-import { App } from './App';
+import { store } from './redux/store';
+import { App } from './components/App';
 import { theme } from 'theme';
 import { ThemeProvider } from 'styled-components';
 import { BrowserRouter } from 'react-router-dom';
