+++ src/pages/DiaryPage/DiaryPage.styles.js
@@ -1,0 +1,54 @@
+import styled from 'styled-components';
+
+export const Date = styled.div`
+  font-size: 34px;
+  font-weight: 700;
+  margin: 100px 15px 0px;
+`;
+
+export const Form = styled.form`
+  display: flex;
+  margin-top: 50px;
+`;
+
+export const Section = styled.div`
+  display: flex;
+  align-items: flex-end;
+`;
+
+export const ProductInput = styled.textarea`
+  font-size: 14px;
+  font-weight: 700;
+  font-family: verdana;
+  width: 240px;
+  height: 37px;
+  border: none;
+  border-bottom: 1px solid #e0e0e0;
+  padding: 0px 0px 12px;
+  margin-left: 15px;
+  resize: none;
+`;
+
+export const GramsInput = styled.input`
+  font-size: 14px;
+  font-weight: 700;
+  font-family: verdana;
+  width: 107px;
+  height: 37px;
+  border: none;
+  border-bottom: 1px solid #e0e0e0;
+  padding: 0px 0px 12px;
+  margin-left: 51px;
+`;
+
+export const Button = styled.button`
+  width: 48px;
+  height: 48px;
+  border-radius: 50%;
+  border: none;
+  background-color: #fc842d;
+  margin-left: 58px;
+  color: white;
+  font-size: 30px;
+  box-shadow: 0px 4px 10px 0px #fc842d;
+`;
