+++ src/pages/LoginPage/LoginPage.js
@@ -1,0 +1,7 @@
+
+const LoginPage = () => {
+    return (<h1>login</h1>);
+
+}
+
+export default LoginPage;
