+++ src/pages/SignupPage/SignupPage.js
@@ -1,0 +1,7 @@
+
+const SignupPage = () => {
+    return (<h1>signup</h1>);
+
+}
+
+export default SignupPage;
