+++ src/redux/auth/operations.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { createAsyncThunk } from '@reduxjs/toolkit';
 
-//axios.defaults.baseURL = process.env.REACT_APP_BASE_URL;
+axios.defaults.baseURL = "http://localhost:3001/api/";
 // axios.defaults.withCredentials = true;
 
 const token = {
