--- src/redux/loader/selectors.js
@@ -1,1 +1,0 @@
-export const selectIsLoading = state => state.loading.isLoading;
